requirements.txt
APScheduler==3.6.3
asgiref==3.5.0
async-lru==1.0.2
attrs==21.2.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==36.0.1
defusedxml==0.7.1
Django==2.2.16
django-debug-toolbar==2.2
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-yasg==1.20.0
Faker==8.13.2
flake8==4.0.1
idna==2.8
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
mixer==7.1.2
more-itertools==8.9.0
mypy==0.910
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.0
pluggy==0.13.1
py==1.10.0
pyaes==1.6.1
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==2.4.7
Pyrogram==1.2.9
PySocks==1.7.1
pytest==5.3.5
python-dateutil==2.8.2
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.1
requests==2.22.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sorl-thumbnail==12.6.3
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tornado==6.1
typing-extensions==3.10.0.2
tzlocal==3.0
uritemplate==4.1.1
urllib3==1.25.11
wcwidth==0.2.5
